refactor: change layout
