+++ src/App.css
@@ -11,7 +11,7 @@
   background-color: #D3D3A3;
 }
 
-.main {
+/* .main {
   display: flex;
   margin: 10px;
-}
+} */
