+++ src/MainPage.css
@@ -1,0 +1,4 @@
+.main {
+  display: flex;
+  margin: 10px;
+}
