+++ src/pages/404/404.css
@@ -1,0 +1,39 @@
+.header {
+  width: 100%;
+  padding: 20px 20px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  gap: 20px;
+  background: linear-gradient(330deg, #AB4F66, #494C77);
+}
+
+.page-title {
+  width: 220px;
+  height: 70px;
+  padding-top: 5px;
+  border: solid 2px #FFFFFF;
+  border-radius: 35px;
+  background-image: url(../../assets/unsplash_wordmark_logo.svg);
+  background-position: bottom center;
+  background-size: 80%;
+  background-repeat: no-repeat;
+  text-align: center;
+}
+
+.page-title span {
+  text-align: center;
+  font-family: Roboto, sans-serif;
+  font-size: 16px;
+  line-height: 1;
+  color: #FFFFFF;
+}
+
+.h2 {
+  width: calc(100% - 250px);
+  text-align: center;
+  font-family: Roboto, sans-serif;
+  font-size: 20px;
+  line-height: 1;
+  color: #FFFFFF;
+}
