+++ src/pages/404/404.tsx
@@ -1,0 +1,14 @@
+import "./404.css";
+
+export default function Page404() {
+  return (
+    <>
+      <header className="header">
+        <div className="page-title">
+          <span>FIND IMAGES ON</span>
+        </div>
+        <h2 className="h2">PAGE NOT FOUND!</h2>
+      </header>
+    </>
+  );
+}
